requirements.txt
pyrogram
tgcrypto
stripe
pymongo
bs4
pyromod
httpx
speedtest-cli
psutil
httpcore
cryptography
pytz
openai
pycountry
httpx==0.23.0
user_agent
requests_toolbelt
getuseragent
fake-useragent==0.1.11
colorama
faker